StaticTools

Documentation for StaticTools.

StaticTools.SplitMix64

StaticTools.Xoshiro256✴︎✴︎

StaticTools.getc

StaticTools.getchar

StaticTools.newline

StaticTools.parsedlm

StaticTools.putchar

StaticTools.splitmix64

StaticTools.static_rng

StaticTools.stderrp

StaticTools.stdinp

StaticTools.stdoutp

StaticTools.xoshiro256✴︎✴︎

StaticTools.SplitMix64 — Type

SplitMix64([seed::Bits64])

Initialize the internal state of a StaticCompiler-safe (non-allocating) SplitMix64 deterministic pseudorandom number generator, optionally specifying a 64-bit seed (which may be a Float64, Int64, or UInt64).

If a seed is not specified, StaticTools.time() will be used, which returns the current Unix epoch time in seconds.

See also:

splitmix64, rand

Examples

julia> seed = StaticTools.time() # Pick a seed
1649890154

julia> rng = SplitMix64(seed) # Initialize the generator
SplitMix64{Int64}((1649890154,))

julia> splitmix64(rng) # Draw a pseudorandom `UInt64` from the generator
0xca764ac7b7ea31e8

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.8704883051360292

StaticTools.Xoshiro256✴︎✴︎ — Type

Xoshiro256✴︎✴︎(seed::NTuple{4,Bits64})

Initialize the internal state of a StaticCompiler-safe (non-allocating) Xoshiro256✴︎✴︎ deterministic pseudorandom number generator, specifying a 256-bit seed, which should be specified as an NTuple of four 64-bit numbers (all either Float64, Int64, or UInt64).

See also:

xoshiro256✴︎✴︎, static_rng, rand

Examples

julia> seed = (0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b)
(0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b)

julia> rng = Xoshiro256✴︎✴︎(seed) # Initialize the generator
Xoshiro256✴︎✴︎{UInt64}((0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b))

julia> xoshiro256✴︎✴︎(rng) # Draw a pseudorandom `UInt64` from the generator
0x11059b6384fba06a

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.9856766307398369

StaticTools.getc — Method

getc(fp::Ptr{FILE})

Libc getc function, accessed by direct llvmcall.

Reads a single character from file pointer fp, returning as Int32. Yields -1 on EOF.

Examples

julia> getc(stdinp())
c
99

StaticTools.getchar — Method

getchar()

Libc getchar function, accessed by direct llvmcall.

Reads a single character from standard input stdin, returning as UInt8.

Examples

julia> getchar()
c
0x63

StaticTools.newline — Method

newline([fp::Ptr{FILE}])

Prints a single newline () to a file pointer fp, defaulting to stdout if not specified. Returns 0 on success.

Examples

julia> putchar('C')
0

julia> newline() # flushes stdout
C
0

StaticTools.parsedlm — Method

parsedlm([T::Type], filepath::String, delimiter::Char)

Parse a delimited text file, given a filepath and delimiter, and return the parsed contents as a MallocMatrix{T}, that is a 2D MallocArray containing numbers of type T.

If not specified, the parse type T will default to Float64.

Examples

julia> using StaticTools

julia> m = (1:10) * (1:10)';

julia> fp = fopen(c"testfile.tsv", c"w"); printf(fp, m); fclose(fp);

julia> parsedlm(Int32, c"testfile.tsv", '	')
10×10 MallocMatrix{Int32}:
  1   2   3   4   5   6   7   8   9   10
  2   4   6   8  10  12  14  16  18   20
  3   6   9  12  15  18  21  24  27   30
  4   8  12  16  20  24  28  32  36   40
  5  10  15  20  25  30  35  40  45   50
  6  12  18  24  30  36  42  48  54   60
  7  14  21  28  35  42  49  56  63   70
  8  16  24  32  40  48  56  64  72   80
  9  18  27  36  45  54  63  72  81   90
 10  20  30  40  50  60  70  80  90  100

StaticTools.putchar — Method

putchar([fp::Ptr{FILE}], c::Union{Char,UInt8})

Libc putchar / fputc function, accessed by direct llvmcall.

Prints a single character c (either a Char or a raw UInt8) to a file pointer fp, defaulting stdout if not specified. Returns 0 on success.

Examples

julia> putchar('C')
0

julia> putchar(0x63)
0

julia> putchar('
') # Newline, flushes stdout
Cc
0

StaticTools.splitmix64 — Function

splitmix64([rng::SplitMix64])

A StaticCompiler-safe (non-allocating) implementation of the SplitMix64 deterministic pseudorandom number generator.

See also:

SplitMix64, rand

Examples

julia> seed = StaticTools.time() # Pick a seed
1649890154

julia> rng = SplitMix64(seed) # Initialize the generator
SplitMix64{Int64}((1649890154,))

julia> splitmix64(rng) # Draw a pseudorandom `UInt64` from the generator
0xca764ac7b7ea31e8

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.8704883051360292

StaticTools.static_rng — Function

static_rng([seed::Bits64])

Initialize a StaticCompiler-safe (non-allocating) deterministic pseudorandom number generator, optionally specifying a 64-bit seed (which may be any 64-bit primitive numeric type – that is, Float64, Int64, or UInt64).

In particular, static_rng uses the specified seed value (or if not specified, the current result of StaticTools.time()) to initialize a simple SplitMix64 generator, which is then in turn used to bootstrap the larger seed required for a Xoshiro256✴︎✴︎ generator.

Examples

julia> rng = static_rng()
Xoshiro256✴︎✴︎{UInt64}((0x2d4c7aa97cc1a621, 0x63460fc58ff25249, 0x81498572d44bd2ec, 0x2d4e96d3a7e9fdd2))

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.6577585429879329

julia> rand(rng)
0.4711097758403277

StaticTools.stderrp — Method

stderrp()

Zero-argument function which returns a raw pointer to the current standard error filestream, stderr.

Examples

julia> stderrp()
Ptr{StaticTools.FILE} @0x00007fffc92b9240

julia> printf(stderrp(), c"Hi there!
")
Hi there!
10

StaticTools.stdinp — Method

stdinp()

Zero-argument function which returns a raw pointer to the current standard input filestream, stdin.

Examples

julia> stdinp()
Ptr{StaticTools.FILE} @0x00007fffc92b9110

StaticTools.stdoutp — Method

stdoutp()

Zero-argument function which returns a raw pointer to the current standard output filestream, stdout.

Examples

julia> stdoutp()
Ptr{StaticTools.FILE} @0x00007fffc92b91a8

julia> printf(stdoutp(), c"Hi there!
")
Hi there!
10

StaticTools.xoshiro256✴︎✴︎ — Method

xoshiro256✴︎✴︎(rng::Xoshiro256✴︎✴︎)

A StaticCompiler-safe (non-allocating) implementation of the Xoshiro256✴︎✴︎ deterministic pseudorandom number generator, written in LLVM IR and invoked via llvmcall.

See also:

Xoshiro256✴︎✴︎, static_rng, rand

Examples

julia> seed = (0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b);

julia> rng = Xoshiro256✴︎✴︎(seed) # Initialize the generator
Xoshiro256✴︎✴︎{UInt64}((0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b))

julia> xoshiro256✴︎✴︎(rng) # Draw a pseudorandom `UInt64` from the generator
0x11059b6384fba06a

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.9856766307398369