StaticTools

Documentation for StaticTools.

StaticTools.MallocArray

StaticTools.MallocArray

StaticTools.MallocMatrix

StaticTools.MallocString

StaticTools.MallocString

StaticTools.MallocString

StaticTools.MallocVector

StaticTools.SplitMix64

StaticTools.StaticString

StaticTools.Xoshiro256✴︎✴︎

Base.parse

StaticTools.fclose

StaticTools.fopen

StaticTools.free

StaticTools.fseek

StaticTools.getc

StaticTools.getchar

StaticTools.gets!

StaticTools.malloc

StaticTools.memcmp

StaticTools.memcpy!

StaticTools.newline

StaticTools.parsedlm

StaticTools.perror

StaticTools.printf

StaticTools.printf

StaticTools.printf

StaticTools.putchar

StaticTools.puts

StaticTools.splitmix64

StaticTools.static_rng

StaticTools.stderrp

StaticTools.stdinp

StaticTools.stdoutp

StaticTools.strlen

StaticTools.strtod

StaticTools.strtol

StaticTools.strtoul

StaticTools.system

StaticTools.time

StaticTools.xoshiro256✴︎✴︎

StaticTools.@c_str

StaticTools.@m_str

StaticTools.MallocArray — Type

MallocArray{T,N} <: AbstractArray{T,N}

N-dimensional dense heap-allocated array with elements of type T.

Much like Base.Array, except (1) backed by memory that is not tracked by the Julia garbage collector (is directly allocated with malloc) so is StaticCompiler-safe, (2) should be freed when no longer in use, and (3) indexing returns views rather than copies.

StaticTools.MallocArray — Method

MallocArray{T}(undef, dims)
MallocArray{T,N}(undef, dims)

Construct an uninitialized N-dimensional MallocArray containing elements of type T. N can either be supplied explicitly, as in Array{T,N}(undef, dims), or be determined by the length or number of dims. dims may be a tuple or a series of integer arguments corresponding to the lengths in each dimension. If the rank N is supplied explicitly, then it must match the length or number of dims. Here undef is the UndefInitializer.

Examples

julia> A = MallocArray{Float64}(undef, 3,3) # implicit N
3×3 MallocMatrix{Float64}:
 3.10504e231   6.95015e-310   2.12358e-314
 1.73061e-77   6.95015e-310   5.56271e-309
 6.95015e-310  0.0           -1.29074e-231

julia> free(A)
0

julia> A = MallocArray{Float64, 3}(undef, 2,2,2) # explicit N
2×2×2 MallocArray{Float64, 3}:
[:, :, 1] =
 3.10504e231  2.0e-323
 2.32036e77   6.94996e-310

[:, :, 2] =
 6.95322e-310  5.0e-324
 6.95322e-310  5.56271e-309

julia> free(A)
0

StaticTools.MallocMatrix — Type

MallocMatrix{T} <: AbstractMatrix{T}

Two-dimensional dense heap-allocated array with elements of type T. As Base.Matrix is to Base.Array, but with MallocArray.

StaticTools.MallocString — Type

struct MallocString
    pointer::Ptr{UInt8}
    length::Int
end

A stringy object that contains length bytes (i.e., UInt8s), including the final null-termination (0x00), at a location in memory specified by pointer.

A MallocString should generally behave like a base Julia String, but is explicitly null-terminated, mutable, standalone-StaticCompiler-safe (does not require libjulia) and backed by malloced memory which is not tracked by the GC and should be freed when no longer in use.

Can be constructed with the m"..." string macro.

StaticTools.MallocString — Method

MallocString(undef, N)

Construct an uninitialized N-byte (including null-termination!) MallocString. Here undef is the UndefInitializer.

Examples

julia> s = MallocString(undef, 10)
m""

julia> free(s)
0

StaticTools.MallocString — Method

MallocString(data::NTuple{N, UInt8})

Construct a MallocString containing the N bytes specified by data.

Examples

julia> data = (0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x20, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x21, 0x00);

julia> s = MallocString(data)
m"Hello world!"

julia> s[8:12] = c"there"; s
m"Hello there!"

julia> free(s)
0

StaticTools.MallocVector — Type

MallocVector{T} <: AbstractVector{T}

Two-dimensional dense heap-allocated array with elements of type T. As Base.Vector is to Base.Array, but with MallocArray.

StaticTools.SplitMix64 — Type

SplitMix64([seed::Bits64])

Initialize the internal state of a StaticCompiler-safe (non-allocating) SplitMix64 deterministic pseudorandom number generator, optionally specifying a 64-bit seed (which may be a Float64, Int64, or UInt64).

If a seed is not specified, StaticTools.time() will be used, which returns the current Unix epoch time in seconds.

See also:

splitmix64, rand

Examples

julia> seed = StaticTools.time() # Pick a seed
1649890154

julia> rng = SplitMix64(seed) # Initialize the generator
SplitMix64{Int64}((1649890154,))

julia> splitmix64(rng) # Draw a pseudorandom `UInt64` from the generator
0xca764ac7b7ea31e8

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.8704883051360292

StaticTools.StaticString — Type

StaticString{N}

A stringy type which should generally behave like a base Julia String, but is explicitly null-terminated, mutable, and standalone-StaticCompiler safe (does not require libjulia).

Can be constructed with the c"..." string macro.

StaticString(data::NTuple{N,UInt8})

Construct a StaticString containing the N bytes specified by data.

Examples

julia> data = (0x48, 0x65, 0x6c, 0x6c, 0x6f, 0x20, 0x77, 0x6f, 0x72, 0x6c, 0x64, 0x21, 0x00);

julia> s = StaticString(data)
c"Hello world!"

julia> s[8:12] = c"there"; s
c"Hello there!"

StaticTools.Xoshiro256✴︎✴︎ — Type

Xoshiro256✴︎✴︎(seed::NTuple{4,Bits64})

Initialize the internal state of a StaticCompiler-safe (non-allocating) Xoshiro256✴︎✴︎ deterministic pseudorandom number generator, specifying a 256-bit seed, which should be specified as an NTuple of four 64-bit numbers (all either Float64, Int64, or UInt64).

See also:

xoshiro256✴︎✴︎, static_rng, rand

Examples

julia> seed = (0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b)
(0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b)

julia> rng = Xoshiro256✴︎✴︎(seed) # Initialize the generator
Xoshiro256✴︎✴︎{UInt64}((0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b))

julia> xoshiro256✴︎✴︎(rng) # Draw a pseudorandom `UInt64` from the generator
0x11059b6384fba06a

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.9856766307398369

Base.parse — Method

parse(::Type{T}, s::Union{StaticString, MallocString})

Parse a number from a StaticString or MallocString s.

Examples

julia> parse(Float64, c"3.141592")
3.141592

julia> parse(Int64, c"3.141592")
3

StaticTools.fclose — Method

fclose(fp::Ptr{FILE})

Libc fclose function, accessed by direct llvmcall.

Closes a file that has been previously opened by fopen, given a file pointer.

See also: fopen, fseek

Examples

julia> fp = fopen(c"testfile.txt", c"w")
Ptr{StaticTools.FILE} @0x00007fffc92bd0b0

julia> printf(fp, c"Here is a string")
16

julia> fclose(fp)
0

shell> cat testfile.txt
Here is a string

StaticTools.fopen — Method

fopen(name, mode)

Libc fopen function, accessed by direct llvmcall.

Returns a file pointer to a file at location specified by name opened for reading, writing, or both as specified by mode. Valid modes include:

c"r": Read, from an existing file.

c"w": Write. If the file exists, it will be overwritten.

c"a": Append, to the end of an existing file.

as well as "r+", c"w+", and "a+", which enable both reading and writing.

See also: fclose, fseek

Examples

julia> fp = fopen(c"testfile.txt", c"w")
Ptr{StaticTools.FILE} @0x00007fffc92bd0b0

julia> printf(fp, c"Here is a string")
16

julia> fclose(fp)
0

shell> cat testfile.txt
Here is a string

StaticTools.free — Method

free(ptr::Ptr)

Libc free function, accessed by direct StaticCompiler-safe llvmcall.

Free memory that has been previously allocated with malloc.

See also: free.

Examples

julia> p = malloc(500)
Ptr{UInt8} @0x00007ff0e9e74290

julia> free(p)
0

StaticTools.fseek — Function

fseek(fp::Ptr{FILE}, offset::Int64, whence::Int32=SEEK_CUR)

Libc fseek function, accessed by direct llvmcall.

Move position within a file given a file pointer fp obtained from fopen. The new position will be offset bytes (or characters, in the event that all characters are non-unicode ASCII characters encoded as UInt8s) away from the position specified by whence.

The position reference whence can take on values of either:

SEEK_SET = Int32(0) File start

SEEK_CUR = Int32(1) Current position

SEEK_END = Int32(2) File end

where SEEK_CUR is the default value.

See also: fopen, fclose

Examples

julia> fp = fopen(c"testfile.txt", c"w+")
Ptr{StaticTools.FILE} @0x00007fffc92bd148

julia> printf(fp, c"Here is a string!")
17

julia> fseek(fp, -2)
0

julia> Char(getc(fp))
'g': ASCII/Unicode U+0067 (category Ll: Letter, lowercase)

julia> fclose(fp)
0

StaticTools.getc — Method

getc(fp::Ptr{FILE})

Libc getc function, accessed by direct llvmcall.

Reads a single character from file pointer fp, returning as Int32 (-1 on EOF).

Examples

julia> getc(stdinp())
c
99

StaticTools.getchar — Method

getchar()

Libc getchar function, accessed by direct llvmcall.

Reads a single character from standard input stdin, returning as UInt8.

Examples

julia> getchar()
c
0x63

StaticTools.gets! — Function

gets!(s::MallocString, fp::Ptr{FILE}, n::Integer=length(s))

Libc fgets function, accessed by direct llvmcall.

Read up to n characters from the filestream specified by file pointer fp to the MallocString s.

Examples

julia> s = MallocString(undef, 100)
m""

julia> gets!(s, stdinp(), 3)
Ptr{UInt8} @0x00007fb15afce550

julia> s
m"
"

StaticTools.malloc — Method

malloc(size::Integer)

Libc malloc function, accessed by direct StaticCompiler-safe llvmcall.

Allocate size bytes of memory and return a pointer to that memory.

See also: free.

Examples

julia> p = malloc(500)
Ptr{UInt8} @0x00007ff0e9e74290

julia> free(p)
0

StaticTools.memcmp — Method

memcmp(a, b, n::Int64)

Libc memcmp function, accessed by direct StaticCompiler-safe llvmcall.

Compare the first n bytes of a and b, returning

a positive value if the first n bytes of a are greater than the first n bytes of b

a negative value if the first n bytes of a are less than the first n bytes of b

0 the first n bytes of a are equal to the first n bytes of b

Examples

julia> memcmp(c"foo", c"foo", 3)
0

julia> memcmp(c"foo", c"bar", 3)
4

StaticTools.memcpy! — Method

memcpy!(a, b, n=length(b))

Libc memcpy function, accessed by direct StaticCompiler-safe llvmcall.

Copy n elements from array b to array a.

Examples

julia> a = rand(3)
3-element Vector{Float64}:
 0.8559883493421137
 0.4203692766310769
 0.5728354965961716

julia> StaticTools.memcpy!(a, ones(3))
0

julia> a
3-element Vector{Float64}:
 1.0
 1.0
 1.0

StaticTools.newline — Method

newline([fp::Ptr{FILE}])

Prints a single newline () to a file pointer fp, defaulting to stdout if not specified.

Returns 0 on success.

Examples

julia> putchar('C')
0

julia> newline() # flushes stdout
C
0

StaticTools.parsedlm — Method

parsedlm([T::Type], filepath::String, delimiter::Char)

Parse a delimited text file, given a filepath and delimiter, and return the parsed contents as a MallocMatrix{T}, that is a 2D MallocArray containing numbers of type T.

If not specified, the parse type T will default to Float64.

Examples

julia> using StaticTools

julia> m = (1:10) * (1:10)';

julia> fp = fopen(c"testfile.tsv", c"w"); printf(fp, m); fclose(fp);

julia> parsedlm(Int32, c"testfile.tsv", '	')
10×10 MallocMatrix{Int32}:
  1   2   3   4   5   6   7   8   9   10
  2   4   6   8  10  12  14  16  18   20
  3   6   9  12  15  18  21  24  27   30
  4   8  12  16  20  24  28  32  36   40
  5  10  15  20  25  30  35  40  45   50
  6  12  18  24  30  36  42  48  54   60
  7  14  21  28  35  42  49  56  63   70
  8  16  24  32  40  48  56  64  72   80
  9  18  27  36  45  54  63  72  81   90
 10  20  30  40  50  60  70  80  90  100

StaticTools.perror — Method

perror(s)

Print the string s to the standard error filestream, stderr.

Returns 0 on success.

Examples

julia> StaticTools.perror(c"ERROR: could not do thing
")
ERROR: could not do thing
0

StaticTools.printf — Method

printf([fp::Ptr{FILE}], [fmt], s)

Libc printf function, accessed by direct llvmcall.

Prints a string s (specified either as a raw Ptr{UInt8} to a valid null-terminated string in memory or else a string type such as StaticString or MallocString for which a valid pointer can be obtained) to a filestream specified by the file pointer fp, defaulting to the current standard output stdout if not specified.

Optionally, a C-style format specifier string fmt may be provided as well.

Returns the number of characters printed on success.

Examples

julia> printf(c"Hello there!
")
Hello there!
13

StaticTools.printf — Method

printf([fp::Ptr{FILE}], [fmt], n::Number)

Libc printf function, accessed by direct llvmcall.

Prints a number n to a filestream specified by the file pointer fp, defaulting to the current standard output stdout if not specified.

Optionally, a C-style format specifier string fmt may be provided as well.

Returns 0 on success.

Examples

julia> printf(1)
1
0

julia> printf(1/3)
3.333333e-01
0

julia> printf(c"%f
", 1/3)
0.333333
0

StaticTools.printf — Method

printf([fp::Ptr{FILE}], a::AbstractArray{<:Number})

Print a matrix or vector of numbers a to a filestream specified by the file pointer fp, defaulting to the current standard output stdout if not specified.

Returns 0 on success.

Examples

julia> printf(rand(5,5))
5.500186e-02    8.425572e-01    3.871220e-01    5.442254e-01    5.990694e-02
5.848425e-01    6.714915e-01    5.616896e-01    6.668248e-01    2.643873e-01
9.156712e-01    1.276033e-01    3.350369e-01    6.513146e-01    9.999104e-01
3.301038e-01    6.027120e-01    5.139433e-01    2.219796e-01    4.057417e-01
2.821340e-01    9.258760e-01    7.950481e-01    1.152236e-01    7.949463e-01
0

StaticTools.putchar — Method

putchar([fp::Ptr{FILE}], c::Union{Char,UInt8})

Libc putchar / fputc function, accessed by direct llvmcall.

Prints a single character c (either a Char or a raw UInt8) to a file pointer fp, defaulting to the current standard output stdout if not specified.

Returns 0 on success.

Examples

julia> putchar('C')
0

julia> putchar(0x63)
0

julia> putchar('
') # Newline, flushes stdout
Cc
0

StaticTools.puts — Method

puts([fp::Ptr{FILE}], s)

Libc puts/fputs function, accessed by direct llvmcall.

Prints a string s (specified either as a raw Ptr{UInt8} to a valid null-terminated string in memory or elseor else a string type such as StaticString or MallocString for which a valid pointer can be obtained) followed by a newline () to a filestream specified by the file pointer fp, defaulting to the current standard output stdout if not specified.

Returns 0 on success.

Examples

julia> puts(c"Hello there!")
Hello there!
0

StaticTools.splitmix64 — Function

splitmix64([rng::SplitMix64])

A StaticCompiler-safe (non-allocating) implementation of the SplitMix64 deterministic pseudorandom number generator.

See also:

SplitMix64, rand

Examples

julia> seed = StaticTools.time() # Pick a seed
1649890154

julia> rng = SplitMix64(seed) # Initialize the generator
SplitMix64{Int64}((1649890154,))

julia> splitmix64(rng) # Draw a pseudorandom `UInt64` from the generator
0xca764ac7b7ea31e8

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.8704883051360292

StaticTools.static_rng — Function

static_rng([seed::Bits64])

Initialize a StaticCompiler-safe (non-allocating) deterministic pseudorandom number generator, optionally specifying a 64-bit seed (which may be any 64-bit primitive numeric type – that is, Float64, Int64, or UInt64).

In particular, static_rng uses the specified seed value (or if not specified, the current result of StaticTools.time()) to initialize a simple SplitMix64 generator, which is then in turn used to bootstrap the larger seed required for a Xoshiro256✴︎✴︎ generator.

Examples

julia> rng = static_rng()
Xoshiro256✴︎✴︎{UInt64}((0x2d4c7aa97cc1a621, 0x63460fc58ff25249, 0x81498572d44bd2ec, 0x2d4e96d3a7e9fdd2))

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.6577585429879329

julia> rand(rng)
0.4711097758403277

StaticTools.stderrp — Method

stderrp()

Zero-argument function which returns a raw pointer to the current standard error filestream, stderr.

Examples

julia> stderrp()
Ptr{StaticTools.FILE} @0x00007fffc92b9240

julia> printf(stderrp(), c"Hi there!
")
Hi there!
10

StaticTools.stdinp — Method

stdinp()

Zero-argument function which returns a raw pointer to the current standard input filestream, stdin.

Examples

julia> stdinp()
Ptr{StaticTools.FILE} @0x00007fffc92b9110

StaticTools.stdoutp — Method

stdoutp()

Zero-argument function which returns a raw pointer to the current standard output filestream, stdout.

Examples

julia> stdoutp()
Ptr{StaticTools.FILE} @0x00007fffc92b91a8

julia> printf(stdoutp(), c"Hi there!
")
Hi there!
10

StaticTools.strlen — Method

strlen(s)

Libc strlen function, accessed by direct StaticCompiler-safe llvmcall.

Returns the length in bytes of the null-terminated string s, not counting the terminating null character.

Examples

julia> strlen("foo") # Not documented, but Julia strings are null-terminated in practice every time I've checked
3

julia> strlen(c"foo")
3

StaticTools.strtod — Method

strtod(s)

Libc strtod function, accessed by direct StaticCompiler-safe llvmcall.

Returns a Float64 ("double") containing the number written out in decimal form in null-terminated string s.

Examples

julia> num, pbuf = StaticTools.strtod(c"3.1415")
(3.1415, ManualMemory.MemoryBuffer{1, Ptr{UInt8}}((Ptr{UInt8} @0x000000010aeee946,)))

julia> num, pbuf = StaticTools.strtod(c"5")
(5.0, ManualMemory.MemoryBuffer{1, Ptr{UInt8}}((Ptr{UInt8} @0x000000010d8f2bb1,)))

StaticTools.strtol — Method

strtol(s)

Libc strtol function, accessed by direct StaticCompiler-safe llvmcall.

Returns an Int64 ("long") containing the number written out in decimal form in null-terminated string s.

Examples

julia> num, pbuf = StaticTools.strtol(c"3.1415")
(3, ManualMemory.MemoryBuffer{1, Ptr{UInt8}}((Ptr{UInt8} @0x000000010dd827f1,)))

julia> num, pbuf = StaticTools.strtol(c"5")
(5, ManualMemory.MemoryBuffer{1, Ptr{UInt8}}((Ptr{UInt8} @0x000000015dbdda41,)))

StaticTools.strtoul — Method

strtoul(s)

Libc strtol function, accessed by direct StaticCompiler-safe llvmcall.

Returns an UInt64 ("unsigned long") containing the number written out in decimal form in null-terminated string s.

Examples

julia> num, pbuf = StaticTools.strtoul(c"3.1415")
(0x0000000000000003, ManualMemory.MemoryBuffer{1, Ptr{UInt8}}((Ptr{UInt8} @0x000000010d6976a1,)))

julia> num, pbuf = StaticTools.strtoul(c"5")
(0x0000000000000005, ManualMemory.MemoryBuffer{1, Ptr{UInt8}}((Ptr{UInt8} @0x000000015ed45d11,)))

StaticTools.system — Method

system(s)

Libc system function, accessed by direct StaticCompiler-safe llvmcall.

Pass the null-terminated string (or pointer thereto) s to the libc system function for evaluation.

Returns 0 on success.

Examples

julia> StaticTools.system(c"time echo hello")
hello

real    0m0.001s
user    0m0.000s
sys 0m0.000s
0

StaticTools.time — Method

time()

Libc time function, accessed by direct StaticCompiler-safe llvmcall.

Return, as an Int64, the current time in seconds since the beginning of the current Unix epoch on 00:00:00 UTC, January 1, 1970.

Examples

julia> StaticTools.time()
1651105298

StaticTools.xoshiro256✴︎✴︎ — Method

xoshiro256✴︎✴︎(rng::Xoshiro256✴︎✴︎)

A StaticCompiler-safe (non-allocating) implementation of the Xoshiro256✴︎✴︎ deterministic pseudorandom number generator, written in LLVM IR and invoked via llvmcall.

See also:

Xoshiro256✴︎✴︎, static_rng, rand

Examples

julia> seed = (0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b);

julia> rng = Xoshiro256✴︎✴︎(seed) # Initialize the generator
Xoshiro256✴︎✴︎{UInt64}((0x9b134eccd2e63538, 0xd74ab64b2c3ecc9b, 0x70ba9c07628c27bf, 0x270a2eb658e6130b))

julia> xoshiro256✴︎✴︎(rng) # Draw a pseudorandom `UInt64` from the generator
0x11059b6384fba06a

julia> rand(rng) # Draw a `Float64` between 0 and 1
0.9856766307398369

StaticTools.@c_str — Macro

@c_str -> StaticString

Construct a StaticString, such as c"Foo".

A StaticString should generally behave like a base Julia String, but is explicitly null-terminated, mutable, and standalone-StaticCompiler safe (does not require libjulia).

Examples

julia> c"Hello there!"
c"Hello there!"

julia> c"foo" == "foo"
true

StaticTools.@m_str — Macro

@m_str -> MallocString

Construct a MallocString, such as m"Foo".

A MallocString should generally behave like a base Julia String, but is explicitly null-terminated, mutable, standalone-StaticCompiler-safe (does not require libjulia) and is backed by mallocd memory which is not tracked by the GC and should be freed when no longer in use.

Examples

julia> s = m"Hello there!"
m"Hello there!"

julia> s == "Hello there!"
true

julia> free(s)
0